requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
cryptography==42.0.5
python-dotenv==1.0.1
aiosqlite==0.19.0
aiohttp==3.9.3
